unify logging functionality
