+++ modules/api-server/src/server/app.ts
@@ -3,7 +3,6 @@
 
 import * as express from "express";
 import * as fs from "fs";
-import { logger as LOGGER } from "./logger";
 import * as path from "path";
 import * as bodyParser from "body-parser";
 import * as expressWinston from "express-winston";
@@ -12,7 +11,7 @@ import * as connector from "./routes/connector";
 import * as events from "./routes/events";
 import * as remoteApis from "./routes/remoteApis";
 import * as localApis from "./routes/localApis";
-import * as configValidator from "@varkes/configuration"
+import { configValidator as cv, logger as lg } from "@varkes/configuration"
 
 const VARKES_LOGO = path.resolve(__dirname, 'resources/logo.svg')
 const LOGO_URL = "/logo";
@@ -22,10 +21,10 @@ const EVENTS_URL = "/events";
 const CONNECTION = "/connection";
 const BATCH_REGISTRATION = "/local/registration";
 const pathToSwaggerUI = require("swagger-ui-dist").absolutePath()
-
+const LOGGER = lg.init("api-server")
 async function init(varkesConfigPath: string, currentPath = "") {
 
-    let varkesConfig = configValidator.discover(LOGGER, varkesConfigPath, currentPath)
+    let varkesConfig = cv.discover(LOGGER, varkesConfigPath, currentPath)
 
     let app = express()
     app.use(bodyParser.json())
@@ -68,4 +67,4 @@ async function init(varkesConfigPath: string, currentPath = "") {
     return app;
 }
 
-export { init }
+export { init, LOGGER }
