+++ modules/api-server/src/server/routes/connector.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 'use strict'
 
-import { logger as LOGGER } from "../logger"
+import { LOGGER } from "../app"
 import * as express from "express"
 import { connection } from "@varkes/app-connector"
