+++ modules/app-connector/src/server/app.ts
@@ -1,7 +1,9 @@
 import * as connection from './connection';
 import { Event } from './event';
 import { API } from './api';
+import { logger as lg } from "@varkes/config-validator"
+const LOGGER: any = lg.init("app-connector")
 connection.init();
 let api = new API();
 let event = new Event();
-export { api, event, connection };
+export { api, event, connection, LOGGER };
