+++ modules/config-validator/src/index.ts
@@ -1,0 +1,4 @@
+import * as configValidator from "./validator"
+import * as logger from "./logger"
+
+export { configValidator, logger }
