+++ modules/odata-mock/package.json
@@ -31,8 +31,7 @@
     "morgan": "^1.9.1",
     "n-odata-server": "^0.4.1",
     "odata2openapi": "^1.3.1",
-    "strong-error-handler": "^3.4.0",
-    "winston": "^3.2.1"
+    "strong-error-handler": "^3.4.0"
   },
   "devDependencies": {
     "@types/mocha": "^5.2.7",
