+++ modules/odata-mock/src/server/app.ts
@@ -5,12 +5,13 @@ const boot = require('loopback-boot');
 const fs = require('fs');
 const express = require('express')
 const bodyParser = require('body-parser');
-import { logger as LOGGER } from "./logger"
 import * as parser from "./parser"
 const path = require("path")
-import * as configValidator from "@varkes/configuration"
+import { configValidator as cv, logger as lg } from "@varkes/configuration"
+
+const LOGGER: any = lg.init("odata-mock")
 async function init(varkesConfigPath: string, currentPath = "") {
-  let varkesConfig = configValidator.discover(LOGGER, varkesConfigPath, currentPath)
+  let varkesConfig = cv.discover(LOGGER, varkesConfigPath, currentPath)
 
   let promises: Promise<any>[] = [];
   for (let i = 0; i < varkesConfig.apis.length; i++) {
@@ -102,4 +103,4 @@ async function generateBootConfig(api: any) {
   return bootConfig
 }
 
-export { init }
+export { init, LOGGER }
