+++ modules/odata-mock/src/server/routes.ts
@@ -3,7 +3,6 @@
 
 const bodyParser = require('body-parser');
 const morgan = require('morgan')
-import { logger as LOGGER } from "./logger"
 const fs = require('fs');
 
 module.exports = function (app: any) {
