+++ modules/openapi-mock/package.json
@@ -18,16 +18,8 @@
     "js-yaml": "^3.13.1",
     "json-to-pretty-yaml": "^1.2.2",
     "morgan": "^1.9.1",
-<<<<<<< HEAD
     "swagger-express-middleware": "^2.0.3",
-    "swagger-ui-dist": "^3.24.0",
-    "winston": "^3.2.1"
-=======
-    "swagger-express-middleware": "^2.0.2",
-    "swagger-ui-dist": "^3.22.3",
-    "winston": "^3.2.1",
-    "@varkes/configuration": "0.0.1"
->>>>>>> rename to configuration
+    "swagger-ui-dist": "^3.24.0"
   },
   "devDependencies": {
     "@types/express": "^4.17.1",
